add verify and fix bugs

+++ scripts/osm_pbf_power_data_extractor.py
@@ -7,6 +7,7 @@
 
 """ OSM extraction script."""
 
+import json
 import logging
 import os
 import pickle
@@ -24,13 +25,13 @@ from esy.osmfilter import osm_info as osm_info
 from esy.osmfilter import osm_pickle as osm_pickle
 from esy.osmfilter import run_filter
 from _helpers import _sets_path_to_root
-from osm_data_config import AFRICA_CC, COMP_CC
+from osm_data_config import AFRICA_CC, COMP_CC, feature_category
 from shapely.geometry import LineString, Point, Polygon
 import hashlib
 
 import logging
-logging.basicConfig()
-_logger=logging.getLogger(__name__)
+# logging.basicConfig()
+_logger=logging.getLogger('osm_data_extractor')
 _logger.setLevel(logging.INFO)
 # logger.setLevel(logging.WARNING)
 
@@ -44,15 +45,9 @@ feature_list = ["substation", "generator", "line", "cable", "tower"]
 # feature_list = ["line"]
 # feature_list = ["substation"]
 
-feature_category = {
-    "substation": "node",
-    "generator": "node",
-    "line": "way",
-    "tower": "node",
-    "cable": "way"
-}
 
-def download_pbf(country_code, update):
+
+def download_pbf(country_code, update, verify):
     """
     Download pbf file from geofabrik for a given country code
 
@@ -86,8 +81,8 @@ def download_pbf(country_code, update):
             with open(PBF_inputfile, "wb") as f:
                 shutil.copyfileobj(r.raw, f)
     
-    if update is True:
-        if verify_pbf(PBF_inputfile, geofabrik_url) is False:
+    if verify is True:
+        if verify_pbf(PBF_inputfile, geofabrik_url, update) is False:
             _logger.warning(f"md5 mismatch, deleting {geofabrik_filename}")
             if os.path.exists(PBF_inputfile):
                 os.remove(PBF_inputfile)
@@ -101,7 +96,7 @@ def download_pbf(country_code, update):
 
 
 verified_pbf =[]
-def verify_pbf(PBF_inputfile, geofabrik_url):
+def verify_pbf(PBF_inputfile, geofabrik_url, update):
     if PBF_inputfile in verified_pbf:
         return True
 
@@ -115,9 +110,10 @@ def verify_pbf(PBF_inputfile, geofabrik_url):
                 hash_md5.update(chunk)
         return hash_md5.hexdigest()
 
-    with requests.get(geofabrik_md5_url, stream=True, verify=False) as r:
-        with open(PBF_md5file, "wb") as f:
-            shutil.copyfileobj(r.raw, f)
+    if update is True or not os.path.exists(PBF_md5file):
+        with requests.get(geofabrik_md5_url, stream=True, verify=False) as r:
+            with open(PBF_md5file, "wb") as f:
+                shutil.copyfileobj(r.raw, f)
 
     local_md5 = calculate_md5(PBF_inputfile)
 
@@ -134,7 +130,7 @@ def verify_pbf(PBF_inputfile, geofabrik_url):
 
 
 pre_filtered = []
-def download_and_filter(feature, country_code, update=False):
+def download_and_filter(feature, country_code, update=False, verify=False):
     """
     Download OpenStreetMap raw file for selected tag.
 
@@ -160,7 +156,7 @@ def download_and_filter(feature, country_code, update=False):
         Nested dictionary with all OpenStreetMap keys of specific component.
         Example of lines. See https://wiki.openstreetmap.org/wiki/Tag:power%3Dline
     """
-    PBF_inputfile = download_pbf(country_code, update)
+    PBF_inputfile = download_pbf(country_code, update, verify)
 
     filter_file_exists = False
     # json file for the Data dictionary
@@ -178,7 +174,7 @@ def download_and_filter(feature, country_code, update=False):
     elementname = f"{country_code}_{feature}s"
 
     # Load Previously Pre-Filtered Files
-    if update is False and filter_file_exists is True:
+    if update is False and verify is False and filter_file_exists is True:
         create_elements = False  # Do not create elements again
 
         ElementsDict = {elementname:{}}
@@ -186,15 +182,17 @@ def download_and_filter(feature, country_code, update=False):
 
         new_prefilter_data = False  # Do not pre-filter data again
         # HACKY: esy.osmfilter code to re-create Data.pickle
-        Data = osm_info.ReadJason(JSON_outputfile, verbose="no")
-        DataDict = {"Data": Data}
-        Data = DataDict["Data"]
+        with open(JSON_outputfile,encoding="utf-8") as f:
+            Data = json.load(f)
+        # DataDict = {"Data": Data}
         # osm_pickle.picklesave(
         #     DataDict,
         #     os.path.realpath(
         #         os.path.join(os.getcwd(), os.path.dirname(JSON_outputfile))
         #     ),
         # )
+        # Data = DataDict["Data"]
+
         _logger.info(f"Loading {feature} Pickle for {AFRICA_CC[country_code]}")
         feature_data = Data, Elements
         return feature_data
@@ -216,7 +214,7 @@ def download_and_filter(feature, country_code, update=False):
         Way: {"power": feature_list},
         Relation: {"power": feature_list},
     }  # see https://dlr-ve-esy.gitlab.io/esy-osmfilter/filter.html for filter structures
-    # HACKY: due to esy.osmfilter validation
+
 
     blackfilter = [
         ("", ""),
@@ -243,6 +241,7 @@ def download_and_filter(feature, country_code, update=False):
         multiprocess=True,
     )
 
+    logging.disable(logging.NOTSET) # Re-enable logging as run_filter disables logging.INFO
     _logger.info(f"Pre: {new_prefilter_data}, Elem: {create_elements}, for {feature} in {country_code}")
 
     feature_data = Data, Elements
@@ -417,7 +416,7 @@ def convert_pd_to_gdf_lines(df_way, simplified=False):
 #     return df_way
 
 
-def process_data(update):
+def process_data(update, verify):
 
     # TODO : Remove columns in favour of a set that drops columns
     columns_tower = [
@@ -520,24 +519,25 @@ def process_data(update):
         )  # Generate GeoJson   
 
     SNGM_CC = {"SN-GM": "senegal-and-gambia"}
+    ZW_CC = {"ZW": "zimbabwe"}
     test_CC = {"LY": "libya", "SZ": "swaziland"}
     # AFRICA_CC = {list of all African countries that are imported from script -> iso_countries_codes}        
     for feature in feature_list:
         df_all_feature = pd.DataFrame()
         for country_code in SNGM_CC.keys():
             # replace Africa_CC by test_CC to only download data for one country
-            feature_data = download_and_filter(feature, country_code, update)
+            feature_data = download_and_filter(feature, country_code, update, verify)
 
             df_node, df_way, Data = convert_filtered_data_to_dfs(country_code, feature_data, feature)
 
             if feature_category[feature] == "way":
-                convert_ways_lines(df_way, Data) if not df_way.empty else _logger.info(f"Empty Way Dataframe for {feature} in {country_code}")
+                convert_ways_lines(df_way, Data) if not df_way.empty else _logger.warning(f"Empty Way Dataframe for {feature} in {country_code}")
                 if not df_node.empty:
-                    _logger.warning(f"Node dataframe should be empty for {feature} in {country_code}")
+                    _logger.warning(f"Node dataframe not empty for {feature} in {country_code}")
 
 
             if feature_category[feature] == "node":
-                convert_ways_points(df_way, Data) if not df_way.empty else _logger.info(f"Empty Way Dataframe for {feature} in {country_code}")
+                convert_ways_points(df_way, Data) if not df_way.empty else None
             
             # Add Type Column
             df_node["Type"] = "Node"
@@ -565,4 +565,4 @@ def process_data(update):
 
 
 if __name__ == "__main__":
-    process_data(update=False) # Set update
+    process_data(update=False, verify=False) # Set update # Verify = True checks local md5s and pre-filters data again
